refactor: Consolidate IRI types into single file
Merge ResourceIri, ValueIri, StandoffMappingIri, and StandoffMappingElementIri
into Iris.scala to reduce file fragmentation and improve code organization.
All functionality and validation logic remains unchanged.
